Driving Ship
Ugh, this was a shitshow...

Around noon I was ready to start programming the way the ship will run,
as the first station to build. I figured I'd do a bit of work to get the
feel of that down before moving on. But I realized that I wanted to ship
to, like, spin around when it hit stuff and not just bounce off the
walls. Godot has a Rigid Physics system for exactly this, so I should
just use it!

Read some docs, watched a tutorial over lunch, seemed easy.

Bad. That was hours and hours of frustration ago. The docs are pretty
sparse, some things are implied more than stated, etc.

But the real issue I found came up with those very rotations I was so
interested in. When applying angled thrust, whether from the back or
from the sides, it would do a weird thing where it would start off
drifting (no rotation), then it would have some drifting and some
rotation, and then it would be all rotation and no drifting. This didn't
_feel_ right to me. I feel like it should either be all rotating at the
beginning, or it should make a circle the whole time. This drifty
loop-de-loop didn't seem intuitive.

I dicked around with stuff for a *long* time. I bounced it off stuff to
see how it would react. I tried lots of things. I still don't know much,
but I _think_ the problem is that the physics engine treats "linear
damping", which prevents drift, differently from "angular damping",
which prevents rotation. Not only are they configured differently, but I
think the two systems are kept totally separate. But I don't think
that's the way things work out here. The reason I stop sliding through a
room is because of the friction on my feet. But if I am spinning in a
room the thing that stops me is still the friction on my feet! It's just
that the friction on my left foot is in one direction and the friction
on the right is in the other, which organically prevents rotation.

And generally, when I'm partially rotating and partially drifting the
thing that sets me straight is just that one of my two wings is moving
through the water faster than the other, which naturally produces a
stronger resisting force.

I think the reason for the stupid loop-de-loop is because in the
beginning it really resists rotation, so it slides. But as it starts to
rotate, the thrust is now changing direction over time. But that means
at some critical point it starts having all the thrusts cancel out, but
the energy spent on rotating is kept, so it just spins. But again, it
feels like just stopping rotation as though it always spins about its
middle, and the middle bearing has friction, just didn't seem to produce
real feeling results for me.

So anyway, I eventually just set both damping values to zero and wrote
my own version of friction that slows the left and right sides
differently based on their actual speeds. Seems to work pretty great so
far. I mean, angling the thrust from a fixed point in the back is still
hard to steer, but at least it feels believable to me now. That's
unintuitive for its own reasons, though, so I switched to a different
model where it shifts power between a left engine and a right engine
based on helm. Seems to work pretty good now.

I want it to be a bit squirrely to wrangle. But intuitively squirrely,
not some sloppy mess.

Anyway, lots of work to get it to be something nice, but at least it
steers!