Crap Driving UI
Ok, so, I had some ideas for a driving UI, but my lack of UI skills
burned a lot of time here.

So then I gave up and just did a thing with bars for now. It's nowhere
near what I wanted, but it's enough to move past this and onto the next
thing!

This also involved a tiny refactor to the relationship between player
and ship, but honestly that was required anyway for the upcoming
networking parts, so this is much better in every way.