Fix In-Ship Collisions
When I bumped into stuff inside the ship, things went **crazy**.

This ship bounced all over the place and things got really bad.
That's because my dude was bouncing inside the ship and that was
applying forces, etc.

I first tried setting layers and masks, but that didn't seem to fix
anything. So I was getting nervous. But then I looked at the docs and it
looks like the fix is just simply choosing "false" for the second
parameter, which now means rather than kicking through a RigidBody I
consider it to be static instead, which is exactly what I wanted!

Seems ok now.