To Helm and Back Again
Ok, I can now walk up to the helm and start driving the boat.
Then I can stop and walk around again.

While I was doing this I noticed a bug that I'd seen before but thought
it was a different bug at the time. All physics things are in global
coordinates, so when I was doing my moving before I was doing it in
world coords.

That wasn't a problem the when I could only ever drive or walk, but
never both. In both of those cases it didn't really matter. But now tha
I can start the machine up, put it into a spin, and then walk around I
got real drunk real fast...