Smidge of Item Cleanup
Things are a little weird in item land...

So, I interact with stuff. This could be instantaneous, like a lever or
something, or it could be persistent like the helm.

So in this case I've decided to do a dance. I interact with the helm,
the helm says "When you interact with me you pick me up", then the
"take" method says "Ok, I have taken you", and then it's actually taking
the helm that does the stateful bits.

The reason for this is twofold:
1. It means the player has a stateful knowledge of the helm. Rather than
   it just triggering a state transform, it's instead a persistent part
   of the player
2. It means the helm can only be taken by one player. I don't
   technically require that as-is, but I do make myself undiscoverable,
   which I hope will make it impossible in practice. I stop the hover
   for me (and eventually the other players) and become "invisible" to
   the action system

And when you put 1 and 2 together you are able to "release" the helm
again later, since we remembered that we'd taken it in the first place.
This allows people to pick it up when I'm done.