Rooms!
Alright, I've got a whole big batch of stuff.
I have rooms, and rooms fill up with water. They can also be connected
to other rooms in which case they will pour their water into the other
room until they both have the same amount of water.

And the water flows between rooms at a rate based on their density. No
one will ever really see the logic here, but I think they may feel that
it feels right. Or not! That's not important.

From there, a player (or other stuff) can detect which room they're in,
and from there what the water level is in that room.
Then, in the player's case, we can use that water level to modulate
their speed.

More will happen later in terms of filling and draining rooms, but this
is a great start!