WaterIndicatorGrid and Room Layout
Alright, I've got the rooms laid out now. To save on a bit of
tediousness I've also made a WaterIndicatorGrid that is just a Vector2
and then it duplicates WaterIndicators that many times.

It's a "tool" script, so it works in the editor. Saved me a ton of time
already!

Also, I realized the NodePath for a WaterIndicator to Room connection is
dumb. Now it just goes up the tree until it finds the Room and then
connects itself up fine. Much much faster and easier to use.