Doors!
Ok, so, the door is drawn and can open and close, but doesn't actually
do anything with water yet.

That's ok, though!

I also had to move around the way the interact code works so I can
interact on the run. Otherwise every time I wanted to open a door I'd
have to stop, then interact, then go.

There's one issue, which is that if I close the door on myself I get
pushed tothe edge of the tilemap, which interacts with the ship to whip
my off into space...

But other than that this is fine!