Initial Vacuum and Player Animation Rebuild
Ok, so, I'm working on getting the vacuum in so that I can clean up
water. Doing this meant attaching it to the sprite, and attaching it to
the sprite meant that I had to flip the attachment point when turning to
the left or right, but you can't scale collision shapes.

So... I had to turn turning into an animation. And that doesn't work
when mixing AnimationPlayer and AnimationTree, which meant basically
rebuild the animation system.

So! It's now tree-based. I think it's actually more work than it used to
be, I kinda wish you could expose some drivers on the outside and use
them inside, because as it stands it's pretty hard-coded.

But whatever! The idea here is that I have a facing value which controls
left or right, and a speed that controls animation speed, and now I can
have the attach point flip when I turn.
It was a bit of a mess, but I got there. For now, at least!

So I can pick up the vacuum, but I can't currently do anything with it.
Or put it down. That part comes next!