Drop Item
I've got a new InputMap item that drops things.
I've also built the drop logic for the vacuum so that I can pick it up
and drop it off. I can also still open doors when I'm holding the
vacuum, but can't iteract with the crate or the helm.

Dope!