Vacuum Complete!
Ok, so, I've made a cute little animation with sfx and now we're all
good!

I hit the button and can still open doors and shit, but if I'm not
pointing at a door it'll vacuum up the water and make a cute sound.

But if I'm trying to vacuum a room that has no water I get a different
cute sound.
It's cute.