Door Sounds
I changed some stuff to animations, and now the doors make sounds.

I couldn't find a sound for "closing door" that I liked, but I was able
to "make" one by screwing with the pitch over time.

Seems a bit flaky, but meh. If it does nothing, then the worst it does
is make a slightly weird sound instead of the one I want.

By making it an animation I was also able to get rid of that defer that
kept some errors out of my log. Now it's on the animation thread, so
it's fine!