Added Walk
Sometimes the player gets kinda jammed between two points and it's hard
to get through the door because one move is a bit too far to the left,
and one step to the right is a bit to far to the right.

So this makes it so I can now half-step and get out of trouble.