Fetching FillyPutty
Ok, this is a vertical slice.
The first thing I added when I was building interactions was a crate,
but it didn't actually do anything. Now I added a "Waiting" state that
pauses a player while they do something, and I used it so now when I
interact with the crate it pauses for a sec, and then gives the player a
putty to hold.

The player can put this down and pick it up again, and players can still
go through doors while holding it.

Finally, damage is now something that can be interacted with, but only
when you have the putty.
When you use the putty on the damage, it goes away!

This is _basically_ a game now!