True Damage
Damage is now created when the ship hits something.
If you hit something very hard it will create level 3 damage right away,
and if you hit it softly enough it won't create any damage at all.

But, if you hit an area that's already damaged, then it will increase
the level of damage.
It's a game!