New Launching Setup
I want the version of the game I run here to be the host for easy
testing, but I don't want to screw up the release of the app in such a
way that players don't get a player screen or something if I forgot to
change thing in the project.

So this is my answer! I make my default screen a launcher that figures
out what to run and then runs that.

I'm sure this will evolve over time, but for now this is the test.