Multiple Players
I have a ways to go still, but this is a good first step!
When I start the game as a host it starts me as the purple character.
When I join as client I'm blue.

I've also got the very beginnings of a sync system in place so that new
players that join get the full situation loaded down before joining,
etc.

I obviously have a long way to go, but they can both walk around and
they both see the players in the right places.