Sync Player Movement Animations
This was actually pretty simple, which makes me happy.

Because I've got my animations built as parameters in the tree player,
and because those show up as properties on the animation tree, and
because Godot's multiplayer system already has a way of syncing
properties between players, all I had to do was add the permissions and
then call "rset" instead of "=" and it just works and now everyone knows
what's up and hears the sounds and everything.

The only thing I didn't do here is that I don't send the player's facing
during snapshot and I don't restore it during load. This means until a
player moves there's a 50/50 chance that they're facing the wrong way.

But I don't think that matters... because as soon as they do something
it'll sync right. So I think it's fine for now.