Item Sync
I now put the items into the sync object.

As an important part of that, I also clear out the default items before
syncing the new ones. This means the vacuum and things don't end up
duplicated if a player moved them, or even more importantly it means the
vacuum isn't duplicated if the player picks it up, and then the other
player joins and sees both the default one and the one in the first
player's hands.

There are a few weird things in here that may not even be required, but
felt important to me at the time. One of them is that when the crate
creates a new bit of goo, I can't just have all of the players make up
their own goo and give it to their own local player. I need to make sure
that when that player puts this goo down later, or deletes the goo
later, it's the same goo that gets put down or deleted for everyone
else.

So to make sure of that I have the player who is interacting with the
crate be the only one that actually makes the goo, and then it syncs
that to all of the players with a name and everything. That way
everything is synced up later.

For now my item sync is a little janky because I didn't actually use a
superclass for items, so I just pull the data out in my data manager
rather than asking them. I should fix that.

I also get the held item for players now. Those live in the player sync
and not the item sync, though.