Sync Damage
We now snapshot where the damage is located and sync it to everyone on
startup. We also make it so only the host notices and calculates damage,
and then they just send that out to everyone else. Due to the fiddlyness
of physics there's no guarantee that two players will have _exactly_ the
same state on whichever frame they each happen to notice the collision,
so they may get different answers about the magnitude or even position
of the damage!

So we just take all of that away and just say only the host does it, and
they'll tell everyone what they decided.