Sync Rooms and Doors
I can now have a player join a game and it will reflect the internal
status of the ship. The doors that are open will be open, the rooms that
are full of water will fill, and the water will flow from the full rooms
to the empty ones just like on the host!

Basically magic...