Sync Phusics to Clients
This means that when a client connects they will see the ship in the
right place and traveling at the right speed.

Just with this simple setup it's a bit choppy, but actually works much
better than I expected!

So I'm going to ship this and the refine it over time.