Sync Helm and Throttle
Ok, there we go!

I now have it so a player gets the helm and throttle when they connect.
This also makese it so a player who isn't the host can drive, and the
host will take their inputs into effect when moving.

It's probably overkill, but I have a sync point here with the helm,
where the player sends all pilot commands to the host, and the host is
the one that tells everyone what the helm and throttle both are.

This way there's less chance that a single message will go missing in
the middle and one player will miss a helm+1 and then just have totally
different physics than everyone else for the rest of the game, etc etc.

With this we can be sure that whatever the inputs are, the result is
unambiguous.