Player Resumption
I now support a player disappearing and coming back gracefully.

When the host sees a player disconnect it tells everyone else that this
player is a piece of garbage who is mean to us, and then everyone starts
fading that player out to indicate that they're not really there.

They also drop whatever they're holding so a disappearing player isn't
holding a critical tool or something.
If they get to the end of fading out, we remove them.

If a player with the same name shows up, we assume it's the same player
but with a different network ID so we just change it to work for them.

On the other side, if I'm joining a game and there was already a player
reported to me that has the same name as me, I assume that's me!

I took the position out of the "player_ready" event thing because it was
kinda meaningless.
Either the player is new so I want to put them in the new player spot,
or they're coming back in which case I want them to be where they
already were...

But for the returning player I'd have to go find my position out of what
I was just told just to report it back to everyone later. Which is dumb.

So I took it out.

There is one edge case here, though. Because I don't sync how long a
player's been mourned, if I join a game where there's an inactive player
there will be a window where the host already deleted them but they
haven't timed out on my computer yet. If the player rejoins in that
time, the host will see them as a new player, and put them in the new
player space. And that "new" player will see that and put themselves in
the new player space. But I will see them wherever they used to be and
resurrect them over there!

But that doesn't really matter, because as soon as the player moved it
would sync their opinion of their position to all clients anyway. So I'd
see it teleport at that time. The only thing that keeping it in the
"player_ready" event thing would do is teleport just before they moved
rather than as soon as they do.

So there's technically a weird thing that could happen where if a player
comes back but doesn't move then different players won't agree on where
they're standing still. Which seems minor.