Fix Helm Sync Bugs
We were treating the helm as an item, but it isn't really one.
Players do take the helm and drop the helm, but they don't _carry_ it
around.

So when syncing I do want to know that they're holding it, but I don't
want to make a new one in the players hands or anything.

The bug here was that "original_layer" wasn't defined for Helm, but
that's actually intentional given that thing I just said. So when
thinking about it I realized the issue is really that it's just a
different kind of thing than the vacuum or the goo.

So I put it in a different spot! And who knows, maybe somewhere in the
future we'll put other stuff there like if the helm can be damaged or
something.

The other bug in there was that the helm moved the player into helm
state when it was taken, but then it was assumed the player would move
themselves back to walking before dropping the helm.

But that was bad because other players synced the "take", but don't run
the input handling code, so they moved into "DRIVING" state on the take
and then never left it, meaning they didn't appear to move anymore, they
just glided. So now we move the player in on take and out on drop, which
is synced and so it's all good now and works with the above.